Feat: setting features

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -11,22 +11,23 @@ use App\Models\User;
 use App\Services\UserService;
 use App\Traits\ApiResponse;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\Log;
+use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
     use ApiResponse;
 
-    public function __construct(protected UserService $userService)
-    {
-        Log::info('UserController constructor called');
-    }
+    public function __construct(protected UserService $userService) {}
 
     /**
      * Display a listing of the resource.
      */
-    public function index(): JsonResponse
+    public function index(Request $request): JsonResponse
     {
+        if (!$request->user()?->can('admin.users.view')) {
+            return $this->forbiddenResponse();
+        }
+
         $users = $this->userService->getAllUsers();
 
         return $this->paginatedResponse(UserResource::collection($users), 'Users retrieved successfully');
@@ -37,6 +38,10 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request): JsonResponse
     {
+        if (!$request->user()?->can('admin.users.add')) {
+            return $this->forbiddenResponse();
+        }
+
         $user = $this->userService->createUser($request->validated());
 
         return $this->createdResponse(UserResource::make($user), 'User created successfully');
@@ -45,8 +50,12 @@ class UserController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(User $user): JsonResponse
+    public function show(Request $request, User $user): JsonResponse
     {
+        if (!$request->user()?->can('admin.users.view')) {
+            return $this->forbiddenResponse();
+        }
+
         return $this->successResponse(UserResource::make($user), 'User retrieved successfully');
     }
 
@@ -55,6 +64,10 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user): JsonResponse
     {
+        if (!$request->user()?->can('admin.users.edit')) {
+            return $this->forbiddenResponse();
+        }
+
         $updatedUser = $this->userService->updateUser($user, $request->validated());
 
         return $this->successResponse(UserResource::make($updatedUser), 'User updated successfully');
@@ -63,9 +76,12 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(User $user): JsonResponse
+    public function destroy(Request $request, User $user): JsonResponse
     {
-        // dd('HERE');
+        if (!$request->user()?->can('admin.users.delete')) {
+            return $this->forbiddenResponse();
+        }
+
         $this->userService->deleteUser($user);
 
         return $this->successResponse(null, 'User deleted successfully');
